Add `libs` option to buildfile, for libraries
Any string passed to the libs option will be added to -l<here> and
passed to the command line during the linking stage. This gets rid
of the iritating "unused command line option" warning when passing
-Wall -Wextra *and* -lXXX to the `flags` option.

This also allows compilers like chibicc to use the buildfile with
libraries, because it doesn't take the -l argument when using -c
for building an object file.